Rename dashlaneImporters to dashlane
Rename the folder
Fix all the imports
Remove dashlaneImporters from white-list
