+++ libs/common/src/services/import.service.ts
@@ -25,8 +25,8 @@ import { ButtercupCsvImporter } from "../importers/buttercup-csv-importer";
 import { ChromeCsvImporter } from "../importers/chrome-csv-importer";
 import { ClipperzHtmlImporter } from "../importers/clipperz-html-importer";
 import { CodebookCsvImporter } from "../importers/codebook-csv-importer";
-import { DashlaneCsvImporter } from "../importers/dashlaneImporters/dashlane-csv-importer";
-import { DashlaneJsonImporter } from "../importers/dashlaneImporters/dashlane-json-importer";
+import { DashlaneCsvImporter } from "../importers/dashlane/dashlane-csv-importer";
+import { DashlaneJsonImporter } from "../importers/dashlane/dashlane-json-importer";
 import { EncryptrCsvImporter } from "../importers/encryptr-csv-importer";
 import { EnpassCsvImporter } from "../importers/enpass-csv-importer";
 import { EnpassJsonImporter } from "../importers/enpass-json-importer";
